Refactor agg strategy
