+++ src/main/kotlin/com/ifpe/edu/br/airpowerserver/controller/DeviceTelemetryController.kt
@@ -32,84 +32,5 @@ class DeviceTelemetryController(private val deviceTelemetryService: DeviceTeleme
         return ResponseEntity.ok(TelemetryAggregationResponse(results = results))
     }
 
-    @PostMapping("/telemetry/aggregate/data-wrapper")
-    fun getDeviceAggregatedDataWrapper(
-        @Valid @RequestBody request: AggregationRequest
-    ): ResponseEntity<AggDataWrapperResponse> {
-        logger.info("Received getDeviceAggregatedDataWrapper request: $request")
 
-        // data wrapper label
-        val aggLabel = "Consumo anual"
-
-        // chard data entries and label
-        val data = ChartDataWrapper(
-            label = "KW/h",
-            listOf(
-                ChatEntry(
-                    "jan",
-                    150
-                ),
-                ChatEntry(
-                    "fev",
-                    250
-                ),
-                ChatEntry(
-                    "mar",
-                    550
-                ),
-                ChatEntry(
-                    "abr",
-                    50
-                ),
-                ChatEntry(
-                    "mai",
-                    140
-                ),
-                ChatEntry(
-                    "jun",
-                    250
-                ),
-                ChatEntry(
-                    "jul",
-                    350
-                ),
-                ChatEntry(
-                    "ago",
-                    0
-                ),
-                ChatEntry(
-                    "set",
-                    0
-                ),
-                ChatEntry(
-                    "out",
-                    0
-                ),
-                ChatEntry(
-                    "nov",
-                    0
-                ),
-                ChatEntry(
-                    "dez",
-                    0
-                )
-            )
-        )
-
-        // aggregated value for given keys and strategy
-        val agg = Agg(
-            label = "Consumo no ano",
-            value = "248,57"
-        )
-
-        // response wrapper
-        return ResponseEntity.ok(
-            AggDataWrapperResponse(
-                label = aggLabel,
-                chartDataWrapper = data,
-                aggregation = agg,
-                size = 15
-            )
-        )
-    }
 }
